requirements.txt
hbutils
hfutils
cheesechaser
dghs-imgutils
numpy
matplotlib
pillow
tqdm
rectpack
torch
torchvision